Rename completion detector modules
